feat: 로그아웃 기능 구현 - #18
