+++ src/components/codeEditer/CodeEditer.js
@@ -17,19 +17,6 @@ function CodeEditer({ socket }) {
 
   ring2.register();
 
-  useEffect(() => {
-    if (Object.keys(runResponse).length === 0) {
-      return;
-    }
-    if (status?.id === 3) {
-      setIsError(false);
-      setResult(decode(stdout || ''));
-    } else if (status?.id > 3) {
-      setIsError(true);
-      setResult(decode(stderr || ''));
-    }
-  }, [runResponse]);
-
   const handleEditorChange = (e) => {
     if (socket) {
       socket.emit('code', e);
@@ -43,6 +30,19 @@ function CodeEditer({ socket }) {
     setIsLoading(false);
     setRunResponse(response);
   };
+
+  useEffect(() => {
+    if (Object.keys(runResponse).length === 0) {
+      return;
+    }
+    if (status?.id === 3) {
+      setIsError(false);
+      setResult(decode(stdout || ''));
+    } else if (status?.id > 3) {
+      setIsError(true);
+      setResult(decode(stderr || ''));
+    }
+  }, [runResponse]);
 
   useEffect(() => {
     if (socket) {
